% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{benchmark}
\alias{benchmark}
\title{benchmark data set}
\format{
A data frame with 11 columns and 150 rows.
}
\source{
mlbench, friedman1
}
\usage{
benchmark
}
\description{
A data frame for regression task generated with mlbench friedman1.
}
\keyword{datasets}
