% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homotypic_cal.R
\name{CalHeteroScore}
\alias{CalHeteroScore}
\title{Calculate the homotypic score of spatial transcriptomics}
\usage{
CalHomoScore(
  spatial_data,
  decon_data,
  n_neigh = 1,
  n_permutation = 50,
  n_cores = 1,
  seed = 123
)
}
\arguments{
\item{spatial_data}{data.frame; coordinates of the spots, using 'row' and 'col' (not 'imagerow' and 'imagecol')}

\item{decon_data}{data.frame; cell type proportions in each spot, with rows as spots and columns as cell types}

\item{n_neigh}{numeric; number of neighboring rings to consider}

\item{n_permutation}{numeric; number of permutations, default is 50}

\item{n_cores}{integer; number of cores for parallel computation, default is 1}

\item{seed}{integer; random seed}
}
\value{
dataframe with spots as rows and homotypic scores of cell types as columns
}
\description{
Calculate the homotypic score of spatial transcriptomics
}
\author{
Jiaxin Luo
}
